🔍 refactor.loading element 정리
