+++ src/components/paymentPage/newPayInput/NewPayInput.tsx
@@ -2,7 +2,7 @@ import * as t from "./newPayInput.style";
 import { useCallback, useState } from "react";
 import { useDaumPostcodePopup } from "react-daum-postcode";
 import { MainButton } from "../../../elements/buttons/Buttons";
-import Input from "../../../elements/Input";
+import Input from "../../../elements/input/Input";
 import { PayMethodSelect } from "../payMethod/PayMethod";
 import { useAppDispatch, useAppSelector } from "../../../redux/store";
 import { RootState } from "../../../redux/store";
