+++ src/components/signUpPage/addressForm/AddressForm.tsx
@@ -1,7 +1,7 @@
 import * as t from "./AddressForm.style";
 import { theme } from "../../../style/theme";
 import { Dispatch, SetStateAction } from "react";
-import Input from "../../../elements/Input";
+import Input from "../../../elements/input/Input";
 import { MainButton } from "../../../elements/buttons/Buttons";
 import { useDaumPostcodePopup } from "react-daum-postcode";
 import { useEffect } from "react";
