+++ src/containers/loginPage/LoginForm.tsx
@@ -5,7 +5,7 @@ import { Link, useNavigate } from "react-router-dom";
 import { openDB } from "idb";
 import { useLoginForm } from "./useLoginForm";
 import { MainButton } from "../../elements/buttons/Buttons";
-import Input from "../../elements/Input";
+import Input from "../../elements/input/Input";
 import SearchModal from "../../components/modal/searchModal/schModal/SearchModal";
 
 const LoginForm = () => {
