+++ src/containers/signupPage/SignUpForm.tsx
@@ -8,7 +8,7 @@ import { SignUpValidation } from "../../shared/utils/yup";
 import { useSignUpForm } from "./useSignUpForm";
 import Profile from "../../components/signUpPage/profile/Profile";
 import AddressForm from "../../components/signUpPage/addressForm/AddressForm";
-import Input from "../../elements/Input";
+import Input from "../../elements/input/Input";
 import { MainButton } from "../../elements/buttons/Buttons";
 
 interface FormValue {
